refact(directory structure): реорганизовать структуру компонент
папка с компонентами замусорена + некоторые компоненты по сути являются страницами и отражают весь их контент, но находятся в components
