--- client/components/VList.vue
@@ -1,19 +1,0 @@
-<template>
-	<div class="v-list">
-		<slot />
-	</div>
-</template>
-
-<script>
-export default {
-	name: 'VList',
-}
-</script>
-
-<style lang="less" scoped>
-
-.v-list * {
-	margin-top: 12px;
-}
-
-</style>
