+++ client/components/common/FormSignupStudent.vue
@@ -48,12 +48,12 @@
 <script>
 import { mapMutations } from 'vuex'
 
-import VButton from '~/components/VButton'
-import VInput from '~/components/VInput'
-import VInputAutocomplete from '~/components/VInputAutocomplete'
+import VButton from '~/components/common/VButton'
+import VInput from '~/components/common/VInput'
+import VInputAutocomplete from '~/components/common/VInputAutocomplete'
 
 export default {
-  name: 'TheFormSingupStudent',
+  name: 'FormSingupStudent',
   components: {
     VButton,
     VInput,
