+++ client/components/common/Logo.vue
