+++ client/layouts/empty.vue
@@ -1,13 +1,19 @@
 <template>
 	<div class="container">
+		<logo class="big" />
 		<Nuxt />
 	</div>
 </template>
 
 <script>
+import Logo from '~/components/common/Logo'
+
 export default {
 	name: 'empty',
-	middleware: 'checkUserIsLoggedOut'
+	components: {
+		Logo,
+	},
+	middleware: 'checkUserIsLoggedOut',
 }
 </script>
