+++ client/pages/bypass-sheets/index.vue
@@ -1,0 +1,23 @@
+<template>
+	<div class="container">
+		<!-- student -->
+		<bypass-sheet-list-student />
+
+		<!-- staff -->
+		<!-- <bypass-sheet-list-staff /> -->
+	</div>
+</template>
+
+<script>
+import bypassSheetListStudent from '~/pages/bypass-sheets/bypassSheetListStudent'
+
+export default {
+	components: {
+		bypassSheetListStudent
+	},
+}
+</script>
+
+<style lang="less" scoped>
+
+</style>
