+++ client/pages/bypass-sheets/signature/_id.vue
@@ -1,0 +1,1 @@
+<!-- Конкретный тип обходного листа -->
