+++ client/pages/bypass-sheets/signature/index.vue
@@ -1,0 +1,1 @@
+<!-- Список типов обходных листов -->
