+++ client/pages/index.vue
@@ -1,7 +1,9 @@
 <template>
 	<div class="container">
+    <!-- Вместо вставки компонент здесь будет редирект на компоненты как на страницы -->
+
 		<!-- student -->
-		<the-page-bypass-sheets />
+		<bypass-sheets />
     <!-- staff -->
 
     <!-- admin -->
@@ -10,12 +12,12 @@
 </template>
 
 <script>
-import ThePageBypassSheets from '~/components/bypass-sheets/ThePageBypassSheets'
+import BypassSheets from '~/pages/bypass-sheets/index'
 
 export default {
   name: 'index',
   components: {
-    ThePageBypassSheets,
+    BypassSheets,
   },
 }
 </script>
